src/main/java/com/nofirst/spring/tdd/sample/sample3/service/TodoService.java
package com.nofirst.spring.tdd.sample.sample3.service;

import com.nofirst.spring.tdd.sample.sample3.domain.Todo;
import com.nofirst.spring.tdd.sample.sample3.repository.TodoRepository;
import java.util.List;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

@Service
public class TodoService {

    private TodoRepository todoRepository;

    @Autowired
    public TodoService(TodoRepository todoRepository) {
        this.todoRepository = todoRepository;
    }

    public List<Todo> findAll() {
        return todoRepository.findAll();
    }

    public Todo save(Todo todo) {
        return todoRepository.save(todo);
    }
}
